refactor: 코드 정리 및 패키지 구조 변경
