+++ src/main/kotlin/calculator/application/parser/Parser.kt
@@ -1,6 +1,6 @@
 package calculator.application.parser
 
-import calculator.common.PositiveInteger
+import calculator.common.model.PositiveInteger
 
 interface Parser {
     fun parse(inputString: String): List<PositiveInteger>
